FeynmanFlow-finance-0.1/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pandas>=2.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
openpyxl>=3.1.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
numpy>=1.26.0 
astrapy>=0.6.0
mysql-connector-python>=8.0.0
scikit-learn>=1.2.0
schedule>=1.2.0
nltk>=3.8
sqlalchemy>=2.0.0